pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = "gooool"
authors = ["zeozeozeo", "dpushin1"]
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
asyncpg = "^0.30.0"
alembic = "^1.14.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
httpx = "^0.28.1"
argon2-cffi = "^23.1.0"
elasticsearch = "^8.17.1"
aiohttp = "^3.11.13"
beautifulsoup4 = "^4.13.3"
minio = "^7.2.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"